Fix bug on apartment and reservation controllers

+++ api/controller/reservation.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, request, jsonify
 from service.reservation import ReservationService
+from service.apartment import ApartmentService
 from service.authentification import AuthentificationService
 from service.user import UserService
 from model.reservation import Reservation
@@ -11,41 +12,58 @@ reservation_blueprint = Blueprint("reservation", __name__)
 
 user_service = UserService()
 service = ReservationService()
+apartment_service = ApartmentService()
 auth = AuthentificationService()
 
 
 @reservation_blueprint.route("/", methods=["POST"])
 def create_reservation():
+    authorization_header = request.headers.get("Authorization")
+    if not authorization_header or not authorization_header.startswith("Basic "):
+        return jsonify({"message": "Authorization header missing or invalid"}), 401
+    admin_username, admin_password = auth.extract_credentials(authorization_header)
+
+    if admin_username == "" or admin_password == "":
+        return jsonify({"message": "Invalid credentials format"}), 401
+
     try:
-        authorization_header = request.headers.get("Authorization")
-        if not authorization_header or not authorization_header.startswith("Basic "):
-            return jsonify({"message": "Authorization header missing or invalid"}), 401
-        username, password = auth.extract_credentials(authorization_header)
-
-        if username == "" or password == "":
-            return jsonify({"message": "Invalid credentials format"}), 401
-
-        if not auth.authenticate_user(username, password):
-            return jsonify({"message": "Invalid username or password"}), 401
-
         req_data = request.get_json()
+        all(
+            key in req_data
+            for key in ("start_date", "end_date", "price", "username", "id_apartment")
+        )
         reservation = Reservation(
             None,
-            req_data["start_date"],
-            req_data["end_date"],
+            None,
+            None,
             req_data["price"],
             req_data["username"],
-            None,
+            req_data["id_apartment"]
         )
+
     except ExServ.ServiceException as e:
         print(e)
         raise ExCon.ControllerException(e.code)
     except Exception as e:
         print(e)
         raise ExCon.ControllerException(400)
-   
+    
+    if not auth.authenticate_admin(admin_username, admin_password):
+        if not auth.authenticate_user(admin_username, admin_password) or admin_username != reservation.username:
+            raise ExCon.ControllerException(401)
+
     if not service.check_values(reservation):
         raise ExCon.ControllerException(400)
+    
+    if not apartment_service.get(reservation.id_apartment):
+        raise ExCon.ControllerException(404)
+
+    try:
+        reservation.start_date = service.string_to_date(req_data["start_date"])
+        reservation.end_date = service.string_to_date(req_data["end_date"])
+    except Exception:
+        raise ExCon.ControllerException(400)
+
     if not user_service.check_user(reservation.username):
         raise ExCon.ControllerException(404)
 
@@ -68,18 +86,25 @@ def get_reservations():
 def get_reservation(reservation_id: int):
     try:
         reservation = service.get(reservation_id)
+        return jsonify(reservation)
     except ExServ.ServiceException as e:
+        print(e)
         raise ExCon.ControllerException(e.code)
-    except Exception:
-        raise ExCon.ControllerException(500)
-    if reservation:
-        return jsonify({"reservation": reservation})
-    else:
-        return jsonify({"message": "Reservation not found"}), 404
+    except Exception as e:
+        print(e)
+        raise ExCon.ControllerException(500)    
+         
 
 
 @reservation_blueprint.route("/<int:reservation_id>", methods=["PATCH"])
 def update_reservation(reservation_id: int):
+    authorization_header = request.headers.get("Authorization")
+    if not authorization_header or not authorization_header.startswith("Basic "):
+        return jsonify({"message": "Authorization header missing or invalid"}), 401
+    admin_username, admin_password = auth.extract_credentials(authorization_header)
+
+    if admin_username == "" or admin_password == "":
+        return jsonify({"message": "Invalid credentials format"}), 401
     try:
         req_data = request.get_json()
         reservation = Reservation(
@@ -95,21 +120,54 @@ def update_reservation(reservation_id: int):
     except Exception:
         raise ExCon.ControllerException(400)
     
-    service.update(reservation)
-    return (
-        jsonify({"message": f"Successfully updated reservation: {reservation_id}"}),
-        200,
-    )
+    if not auth.authenticate_admin(admin_username, admin_password):
+        if not auth.authenticate_user(admin_username, admin_password) or admin_username != reservation.username:
+            raise ExCon.ControllerException(401)
+
+    if not service.check_values(reservation):
+        raise ExCon.ControllerException(400)
+    
+    if not apartment_service.get(reservation.id_apartment):
+        raise ExCon.ControllerException(404)
+
+    try:
+        service.update(reservation)
+        return (
+            jsonify({"message": f"Successfully updated reservation: {reservation_id}"}),
+            200,
+        )
+    except Exception:
+        raise ExCon.ControllerException(500)
 
 
 @reservation_blueprint.route("/<int:reservation_id>", methods=["DELETE"])
 def delete_reservation(reservation_id: int):
+    authorization_header = request.headers.get("Authorization")
+    if not authorization_header or not authorization_header.startswith("Basic "):
+        return jsonify({"message": "Authorization header missing or invalid"}), 401
+    admin_username, admin_password = auth.extract_credentials(authorization_header)
+
+    if admin_username == "" or admin_password == "":
+        return jsonify({"message": "Invalid credentials format"}), 401
     try:
         reservation = service.get(reservation_id)
     except ExServ.ServiceException as e:
         raise ExCon.ControllerException(e.code)
     except Exception:
         raise ExCon.ControllerException(400)
+    
+    reservation_username = service.get_username(reservation_id)
+
+    if not auth.authenticate_admin(admin_username, admin_password):
+        if not auth.authenticate_user(admin_username, admin_password) or admin_username != reservation_username:
+            raise ExCon.ControllerException(401)
+
+    if not service.check_values(reservation):
+        raise ExCon.ControllerException(400)
+    
+    if not apartment_service.get(reservation.id_apartment):
+        raise ExCon.ControllerException(404)
+
     if reservation:
         service.delete(reservation_id)
         return (
